bolos: fix sys_cx_hash_sha256 prototype for 64-bit builds

+++ src/bolos/cx_sha256.c
@@ -248,9 +248,9 @@ int cx_sha256_validate_context(const cx_sha256_t *ctx)
 /* ----------------------------------------------------------------------- */
 /*                                                                         */
 /* ----------------------------------------------------------------------- */
-int sys_cx_hash_sha256(const unsigned char *in, unsigned int len,
+int sys_cx_hash_sha256(const unsigned char *in, size_t len,
                        unsigned char *out,
-                       unsigned int out_len __attribute__((unused)))
+                       size_t out_len __attribute__((unused)))
 {
   cx_sha256_init(&G_cx.sha256);
   cx_sha256_update(&G_cx.sha256, in, len);
